-added rol, rcl. passes testcases except flags. -bitwise.rotateleft()
Had a problem where the intel manual states the only time OF is only set is when the rotation is by 1(probably because there would be no actual usage of it otherwise), however my machine would set the OF on most of the ROL/RCL operations. Couldn't figure out what exactly for, it definitely wasn't the same as for 1 bit rotations, because of this I will leave it how the intel manual describes. I had to ask a question here, https://stackoverflow.com/questions/57747329/inconsistent-behaviour-of-rol-and-of . Turns out my understanding of "undefined" was wrong, I thought, and so far was true, that "undefined" meant that the flag wouldn't be set at all after that operation. Until now, my machine and intel manual have been in agreement but I think in the future I will continue to side with the manual, just so I can be sure of implementation. I'm not going to change anything i've done so far because of this, as I agree with what is said in the answer, there are probably many different conditions on different models.