Many improvements to AddressRanges. The binary search was rewritten and is now simpler, and actually works properly as well as TryMerge() reintroduced.
MemoryListView now takes a memoryspace and formats that rather than passing the data already formatted. It is now much more independent of a control.

Greatly improved the external behaviour of a MemorySpace. It now works with address ranges such that every address stored is accounted for in an address map and is simple to work with the address space externally. See MemoryListView for specifics on implementation of this.

Also lots of comments